refactor: move `config.Upstream` into own file
